% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetchActivities}
\alias{fetchActivities}
\title{Fetches activities from currect cruise}
\usage{
fetchActivities(apiAdress = API_ADRESS)
}
\arguments{
\item{apiAdress}{address to toktlogger API}
}
\value{
table of activities
}
\description{
Fetches activities from currect cruise
}
