% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetchActivity}
\alias{fetchActivity}
\title{Fetches one activity from current cruise}
\usage{
fetchActivity(
  localStationNumber,
  activitygroup = ACTIVITYGROUP_DEFAULT,
  apiAdress = API_ADRESS
)
}
\arguments{
\item{localStationNumber}{local station number for activity}

\item{apiAdress}{address to toktlogger API}

\item{activityMainGroupName}{activity group, see examples for complete list}
}
\value{
activity, one row formatted as the return from \code{\link[toktlogger]{fetchActivities}}
}
\description{
Fetches one activity from current cruise
}
\examples{
 #find activity type names
 activities <- fetchActivities()
 activitygroups <- unique(activities$activityMainGroupName)

 #find valid local station numbers
 locStations <- activities$localstationNumber[activities$activityMainGroupName==activitygroups[1]]

 #find the with activity type
 fetchActivity(locStations[1], activitygroups[1])

}
