% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetchNmeaAtActivity}
\alias{fetchNmeaAtActivity}
\title{Fetches all NMEA telegrams at activity}
\usage{
fetchNmeaAtActivity(
  localStationNumber,
  activitygroup = ACTIVITYGROUP_DEFAULT,
  apiAdress = API_ADRESS
)
}
\arguments{
\item{localStationNumber}{local station number for activity}

\item{apiAdress}{address to toktlogger API}

\item{activityMainGroupName}{activity group, see examples for complete list}
}
\value{
all NMEA sentences at activity
}
\description{
Fetches all NMEA telegrams at activity
}
\examples{
 #for an example of output
 data("NMEAexample")
}
