% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDepthProfile}
\alias{plotDepthProfile}
\title{Plots depth profile}
\usage{
plotDepthProfile(nmeaMessages)
}
\arguments{
\item{nmeaMessages}{nmea message for station, see \code{\link[toktlogger]{extractTrawlSensorData}}}
}
\value{
ggplot2 object
}
\description{
Plots depth profile from trawl door sensors
}
