% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDoorSpread}
\alias{plotDoorSpread}
\title{Plots door spread}
\usage{
plotDoorSpread(nmeaMessages, tolerance = c(48, 52), ylim = c(35, 65))
}
\arguments{
\item{nmeaMessages}{nmea message for station, see \code{\link[toktlogger]{extractTrawlSensorData}}}

\item{tolerance}{upper and lower limit tolerated for door spread}

\item{ylim}{y-axis (door spread) range of plot}
}
\value{
ggplot2 object
}
\description{
Plots door spread
}
