% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotReport}
\alias{plotReport}
\title{Plots trawl calibration report}
\usage{
plotReport(nmeaMessages)
}
\arguments{
\item{nmeaMessages}{nmea message for station, see \code{\link[toktlogger]{extractTrawlSensorData}}}
}
\description{
Plots trawl calibration report
}
