% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTrawlOpening}
\alias{plotTrawlOpening}
\title{Plots trawl opening}
\usage{
plotTrawlOpening(nmeaMessages, tolerance = c(3.5, 4.5), ylim = c(2.5, 5.5))
}
\arguments{
\item{nmeaMessages}{nmea message for station, see \code{\link[toktlogger]{extractTrawlSensorData}}}

\item{tolerance}{upper and lower limit tolerated for trawl opening}

\item{ylim}{y-axis (trawl opening) range of plot}
}
\value{
ggplot2 object
}
\description{
Plots trawl opening
}
