% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{setActvityGroupDefault}
\alias{setActvityGroupDefault}
\title{Set default Activity Main Group Name}
\usage{
setActvityGroupDefault(activityGroup)
}
\arguments{
\item{activityGroup}{default Activity Main Group Name}
}
\description{
Set default Activity Main Group Name
}
